requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.16
django-environ==0.8.1
gunicorn==20.1.0
jalali_core==1.0.0
jdatetime==5.0.0
joblib==1.4.0
numpy==1.26.4
pandas==2.2.2
pycodestyle==2.8.0
python-dateutil==2.9.0.post0
pytz==2021.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.4.0
toml==0.10.2
tzdata==2024.1
whitenoise==5.3.0
xgboost==2.0.3
